Move single order solver settlement merging functions into module
